Add cohesion rule for boids and pseudo code
The cohesion rule, as implemented by the cohesion function, makes a boid
turn towards the center of mass of all boids in its visual range.
This causes an effect that kind of looks like the typical visualization
of an atom.

A side effect of this is that the x and y coordinates of a boid is now
a floating point number. Handling for this during rendering is added.